requirements.txt
Django==1.10
django-bootstrap3==6.2.2
django-crequest==2014.9.19
django-forms-bootstrap
django-extensions
django-fixtureless
django-crispy-forms
django-simple-history
django-autocomplete-light
django-suit
django-tables2
django-pagination
django-bootstrap3-datepicker
django-filter
django-modeltranslation
django-ckeditor
mysqlclient==1.3.10
Pillow
django-geojson
django-leaflet
django-nvd3
django-bower
jsonfield
